A few more missing renames
